rename wasm to promise
