+++ Cargo.toml
@@ -23,7 +23,7 @@ async-scheduler = ["tokio?/time"]
 default = ["uuid"]
 thread = []
 uuid = ["dep:uuid", "uuid/v4"]
-wasm = ["wasm-bindgen-futures"]
+promise = ["wasm-bindgen-futures"]
 tokio = ["dep:tokio", "tokio/sync", "tokio/rt"]
 async-std = ["dep:async-std"]
 cron = ["dep:cron"]
