+++ src/runtime/promise.rs
@@ -4,19 +4,19 @@ use crate::schedule::IntoSchedule;
 use crate::{Runtime, Task};
 use wasm_bindgen_futures::spawn_local;
 
-/// The runtime for WebAssembly.
+/// The runtime for javascript environment.
 ///
 /// the task is spawned using [`wasm_bindgen_futures::spawn_local`], and be executed with promise.
 ///
 /// # Create a new task
 /// see [`Task::promise`]
-pub struct Wasm;
+pub struct Promise;
 
-impl Runtime for Wasm {
+impl Runtime for Promise {
     type Handle = ();
 }
 
-impl Task<Wasm> {
+impl Task<Promise> {
     /// Create a new task that will be executed with promise.
     ///
     /// # Example
