+++ src/schedule/period.rs
@@ -84,23 +84,3 @@ impl IntoSchedule for TimeDelta {
     }
 }
 
-#[test]
-fn test_forward() {
-    use crate::prelude::*;
-    let now = Utc::now();
-    let mut period = Period::new(TimeDelta::days(10), now);
-    period.forward_to(now + TimeDelta::days(7));
-    assert_eq!(
-        chrono::DurationRound::duration_round(period.get_next(), TimeDelta::milliseconds(1))
-            .unwrap(),
-        chrono::DurationRound::duration_round(
-            Utc::now() + TimeDelta::days(10),
-            TimeDelta::milliseconds(1)
-        )
-        .unwrap()
-    );
-
-    let mut period = Period::new(TimeDelta::new(30, 30_123_456).unwrap(), now);
-    period.forward_to(now + TimeDelta::days(1));
-    println!("now:{now}, {:?}", period);
-}
